requirements.txt
Django==1.9.10
django-filter==0.15.2
djangorestframework==3.4.7
Markdown==2.6.7
Pillow==3.3.1
psycopg2==2.6.2
wheel==0.24.0
